160 - Rename 'MetadataTable' to 'Accordion' and make it reusable
