+++ frontend/src/components/metadata/SuiteMetadata.jsx
@@ -1,17 +1,13 @@
 ﻿import React from 'react';
-import MetadataTable from './MetadataTable';
+import Accordion from '../accordion/Accordion';
 import { useStateValue } from '../../contexts/state';
+import { useTranslation } from 'react-i18next';
 
 const SuiteMetadata = () => {
+    const [t] = useTranslation(['accordion']);
     const [{ selectedSuiteState }] = useStateValue();
 
-    const name = [
-        'Suite Id:',
-        'Name:',
-        'Full name:',
-        'Repository:',
-        'Starttime:',
-    ];
+    const name = ['Suite Id', 'Name', 'Full name', 'Repository', 'Starttime'];
 
     const value = selectedSuiteState && [
         selectedSuiteState.suite.id.toString(),
@@ -23,7 +19,11 @@ const SuiteMetadata = () => {
             : '',
     ];
 
-    return selectedSuiteState && <MetadataTable name={name} value={value} />;
+    return (
+        selectedSuiteState && (
+            <Accordion header={t('metadata')} name={name} value={value} />
+        )
+    );
 };
 
 export default SuiteMetadata;
