+++ frontend/src/locales/en/accordion.json
