+++ frontend/src/utils/i118n.js
@@ -9,7 +9,7 @@ import overview from '../locales/en/overview.json';
 import header from '../locales/en/header.json';
 import history from '../locales/en/history.json';
 import buttons from '../locales/en/buttons.json';
-import metadata from '../locales/en/metadata.json';
+import accordion from '../locales/en/accordion.json';
 
 const resources = {
     en: {
@@ -22,7 +22,7 @@ const resources = {
         header,
         history,
         buttons,
-        metadata,
+        accordion,
     },
 };
